Collapse task impl.
